not gonnna get the footer, leaving it for now. Signup form adjusted
