+++ react-app/src/components/EventForm/EventForm.css
@@ -26,7 +26,7 @@
   
   .event-form-submit-button {
     padding: 10px;
-    color: #5DACBD#2C2629;
+    color: #2C2629;
     background-color: #5DACBD;
     border: 2px solid #5DACBD;
     border-radius: 999px;
