+++ react-app/src/components/GroupForm/GroupForm.css
@@ -26,10 +26,11 @@ input {
 
 .group-form-submit-button {
   padding: 10px;
-  color: #5DACBD;
-  background-color: #2C2629;
+  color: #2C2629;
+  background-color: #5DACBD;
   border: 2px solid #5DACBD;
   border-radius: 999px;
+  font-size: 16px;
 }
 
 .group-form-submit-button:hover {
