--- react-app/src/components/auth/SignUpForm.js
@@ -1,112 +1,0 @@
-import React, { useState } from "react";
-import { Redirect } from 'react-router-dom';
-import { signUp } from '../../services/auth';
-
-const SignUpForm = ({authenticated, setAuthenticated}) => {
-  const [username, setUsername] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [repeatPassword, setRepeatPassword] = useState("");
-  const [imageUrl, setImageUrl] = useState("");
-  const [zipCode, setZipCode] = useState(0);
-  const onSignUp = async (e) => {
-    e.preventDefault();
-    if (password === repeatPassword) {
-      const user = await signUp(username, email, password, imageUrl, zipCode);
-      if (!user.errors) {
-        setAuthenticated(true);
-      }
-    }
-  };
-
-  const updateUsername = (e) => {
-    setUsername(e.target.value);
-  };
-
-  const updateEmail = (e) => {
-    setEmail(e.target.value);
-  };
-
-  const updatePassword = (e) => {
-    setPassword(e.target.value);
-  };
-
-  const updateRepeatPassword = (e) => {
-    setRepeatPassword(e.target.value);
-  };
-
-  const updateImageUrl = (e) => {
-    setImageUrl(e.target.value);
-  };
-
-  const updateZipCode = (e) => {
-    setZipCode(e.target.value);
-  };
-
-  if (authenticated) {
-    return <Redirect to="/" />;
-  }
-
-  return (
-    <form onSubmit={onSignUp}>
-      <div>
-        <label>User Name</label>
-        <input
-          type="text"
-          name="username"
-          onChange={updateUsername}
-          value={username}
-        ></input>
-      </div>
-      <div>
-        <label>Email</label>
-        <input
-          type="text"
-          name="email"
-          onChange={updateEmail}
-          value={email}
-        ></input>
-      </div>
-      <div>
-        <label>Profile Photo</label>
-        <input
-          type="text"
-          name="imageUrl"
-          onChange={updateImageUrl}
-          value={imageUrl}
-        ></input>
-      </div>
-      <div>
-        <label>Zip Code</label>
-        <input
-          type="text"
-          name="zipCode"
-          onChange={updateZipCode}
-          value={zipCode}
-        ></input>
-      </div>
-      <div>
-        <label>Password</label>
-        <input
-          type="password"
-          name="password"
-          onChange={updatePassword}
-          value={password}
-        ></input>
-      </div>
-      <div>
-        <label>Repeat Password</label>
-        <input
-          type="password"
-          name="repeat_password"
-          onChange={updateRepeatPassword}
-          value={repeatPassword}
-          required={true}
-        ></input>
-      </div>
-      <button className="authButton" type="submit">Sign Up</button>
-    </form>
-  );
-};
-
-export default SignUpForm;
