+++ react-app/src/index.css
@@ -35,6 +35,3 @@ a {
     text-decoration: none;
     color: #5dacbd;
 }
-
-
-
